feat: add dashboard feature
