+++ src/frontend/src/lib/layout/components/header.tsx
@@ -12,8 +12,8 @@ type NavLinkItem = {
 const navLinks: NavLinkItem[] = [
     { to: "/", label: "Home" },
     { to: "/about", label: "About" },
-    { to: "/votes/history", label: "History", protected: true },
-    { to: "/wallet", label: "Wallet", protected: true }
+    { to: "/dashboard", label: "Dashboard", protected: true  },
+    { to: "/wallet", label: "Wallet", protected: true  },
 ];
 
 export default function Header() {
