+++ src/frontend/src/services/wallet.ts
@@ -1,0 +1,8 @@
+import { Principal } from "@dfinity/principal";
+import { ICPService } from "./icp";
+
+export class WalletService {
+    static async getIcpBalance(principal: string) {
+        return await ICPService.getBalance(Principal.fromText(principal));
+    }
+}
